requirements.txt
supabase==1.2.0
requests==2.31.0
geopandas==0.14.3
fiona==1.9.5
streamlit==1.32.0
pandas==2.2.0
plotly==5.19.0
python-dotenv==1.0.1
openpyxl>=3.1.0
Pillow>=10.0.0
